+++ lib/feature/map/repository/search_using_room.dart
@@ -1,0 +1,1 @@
+// TODO: MapBottomInfo内のSetUsingColorをここに移す
